requirements.txt
fastapi
tortoise-orm
asyncpg
passlib[bcrypt]
python-dotenv
cryptography
uvicorn
